Move CourtCaseDetailsTabs & CourtCaseDetailsPanel into a new folder
